+++ modules/common/default.nix
@@ -7,6 +7,7 @@
     ./user-nixos.nix
     ./users.nix
     ./config-merge.nix
+    ./bootstrap.nix
   ];
 
   # Core Nix Settings